requirements.txt
altair==5.5.0
annotated-types==0.7.0
anyio==4.8.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.8.1
asttokens==3.0.0
async-lru==2.0.5
attrs==25.3.0
audioop-lts==0.2.1
autopep8==2.0.2
babel==2.17.0
beautifulsoup4==4.13.4
binaryornot==0.4.4
bleach==6.2.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.8.3
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.1.0
cli_helpers==2.4.0
click==8.1.8
colorama==0.4.6
comm==0.2.2
comtypes==1.4.10
configobj==5.0.9
cookiecutter==2.6.0
crispy-bootstrap5==2024.10
cryptography==44.0.2
debugpy==1.8.14
decorator==5.2.1
defusedxml==0.7.1
distro==1.9.0
Django==5.2.5
django-bootstrap5==25.1
django-cors-headers==4.7.0
django-crispy-forms==2.3
djangorestframework==3.15.2
dnspython==2.7.0
email_validator==2.2.0
executing==2.2.0
fastapi-cli==0.0.7
fastjsonschema==2.21.1
Flask==3.1.0
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
flet==0.25.2
flet-cli==0.25.2
fqdn==1.5.1
gitdb==4.0.12
GitPython==3.1.45
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.181.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
greenlet==3.2.4
griffe==1.14.0
grpcio==1.74.0
grpcio-status==1.71.2
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.7
httplib2==0.30.0
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.1
idna==3.4
ipykernel==6.29.5
ipython==9.2.0
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.7
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.5
jiter==0.10.0
json5==0.12.0
jsonpointer==3.0.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
jupyter-console==6.6.3
jupyter-events==0.12.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyter_server==2.16.0
jupyter_server_terminals==0.5.3
jupyterlab==4.4.3
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.15
lxml==5.3.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mcp==1.13.1
mdurl==0.1.2
mistune==3.1.3
narwhals==1.48.1
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
notebook==7.4.3
notebook_shim==0.2.4
numpy==2.2.3
oauthlib==3.2.2
openai==1.106.1
openai-agents==0.2.11
outcome==1.3.0.post0
overrides==7.7.0
packaging==24.2
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
pillow==11.0.0
platformdirs==4.3.8
playwright==1.54.0
plotly==6.2.0
pocketsphinx==5.0.4
prometheus_client==0.22.0
prompt_toolkit==3.0.50
proto-plus==1.26.1
protobuf==5.29.5
psutil==7.0.0
pure_eval==0.2.3
pyaes==1.6.1
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
PyAudio==0.2.14
pycodestyle==2.10.0
pycparser==2.22
pydantic==2.11.1
pydantic-settings==2.10.1
pydantic_core==2.33.0
pydeck==0.9.1
pydub==0.25.1
pyee==13.0.0
pyfzf==0.3.1
pygame==2.6.1
Pygments==2.19.1
pyjokes==0.8.3
PyMySQL==1.1.1
pyparsing==3.2.3
pyperclip==1.9.0
pypiwin32==223
pypng==0.20220715.0
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.16.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-json-logger==3.3.0
python-multipart==0.0.20
python-slugify==8.0.4
pyttsx3==2.98
pytz==2025.1
pywin32==311
pywinpty==2.0.15
PyYAML==6.0.2
pyzmq==26.4.0
qrcode==7.4.2
referencing==0.36.2
repath==0.9.0
requests==2.31.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.9.4
rich-toolkit==0.14.0
rpds-py==0.25.1
rsa==4.9.1
selenium==4.35.0
Send2Trash==1.8.3
setuptools==78.1.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
sounddevice==0.5.1
soupsieve==2.7
SpeechRecognition==3.9.0
SQLAlchemy==2.0.36
sqlglot==26.9.0
sqlparse==0.5.3
sse-starlette==3.0.2
stack-data==0.6.3
standard-aifc==3.13.0
standard-chunk==3.13.0
starlette==0.46.1
streamlit==1.48.1
tabulate==0.9.0
tenacity==9.1.2
terminado==0.18.1
text-unidecode==1.3
tinycss2==1.4.0
toml==0.10.2
tomli==2.0.1
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
trio==0.30.0
trio-websocket==0.12.2
typer==0.15.2
types-python-dateutil==2.9.0.20241206
types-requests==2.32.4.20250809
typing-inspection==0.4.0
typing_extensions==4.14.1
tzdata==2025.2
uri-template==1.3.0
uritemplate==4.2.0
urllib3==2.5.0
uv==0.5.18
uvicorn==0.34.0
watchdog==4.0.2
watchfiles==1.0.4
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.1.3
whitenoise==6.9.0
widgetsnbextension==4.0.14
wsproto==1.2.0
